booking_passengers/requirements/common.txt
Django==3.1.5
djangorestframework==3.12.2
django-environ==0.4.5
django-cors-headers==3.7.0
requests==2.26.0
isodate==0.6.0
redis==3.5.3
Unidecode==1.1.1
ecs-logging==1.0.0
elastic-apm==6.1.0
elasticsearch==8.0.0
azure-servicebus==7.8.0
aiohttp==3.8.3
